+++ CHANGELOG.md
@@ -44,11 +44,16 @@ Added `validate_take_ownership()` method so that an SNS is able to add a custom 
 
 Added `is_aliased` field to `get_asset_properties` and `set_asset_properties`.
 
+Batch ids are now stable.  With upcoming changes to support asset updates by proposal,
+having the asset canister not reuse batch ids will make it easier to verify that a particular
+batch has been proposed.
+
 ## Dependencies
 
 ### Frontend canister
 
-- Module hash: 98863747bb8b1366ae5e3c5721bfe08ce6b7480fe4c3864d4fec3d9827255480
+- Module hash: 309d95726d26ac1fa595fb90fed056bac78a681eea4a53759338f1882839e731
+- https://github.com/dfinity/sdk/pull/3015
 - https://github.com/dfinity/sdk/pull/3001
 - https://github.com/dfinity/sdk/pull/2987
 - https://github.com/dfinity/sdk/pull/2982