+++ CHANGELOG.md
@@ -38,6 +38,8 @@ Added the ability to configure the WASM module used for assets canisters through
 
 ### fix: dfx deploy and icx-asset no longer retry on permission failure
 
+### feat: --created_at_time for ledger transfer, create-canister, and top-up
+
 ## Asset Canister
 
 Added `validate_take_ownership()` method so that an SNS is able to add a custom call to `take_ownership()`.