+++ CHANGELOG.md
@@ -38,7 +38,7 @@ Added the ability to configure the WASM module used for assets canisters through
 
 ### fix: dfx deploy and icx-asset no longer retry on permission failure
 
-### feat: --created_at_time for ledger transfer, create-canister, and top-up
+### feat: --created-at-time for the ledger functions: transfer, create-canister, and top-up
 
 ## Asset Canister